+++ secp256kfun/examples/quick_bip340.rs
@@ -1,16 +1,17 @@
+//! A quick and dirty (but working) implementation of BIP340 Schnorr signatures.
 #![allow(non_snake_case)]
 use rand::thread_rng;
 use secp256kfun::{
     g,
     hash::{HashAdd, Tagged},
     marker::*,
-    s, Scalar, XOnly, G,
+    s, Point, Scalar, G,
 };
 use sha2::Sha256;
 
 #[derive(Clone, Debug)]
 pub struct Signature {
-    pub R: XOnly,
+    pub R: Point<EvenY>,
     pub s: Scalar<Public, Zero>,
 }
 
@@ -18,16 +19,16 @@ lazy_static::lazy_static! {
     pub static ref BIP340_CHALLENGE: Sha256 = Sha256::default().tagged(b"BIP0340/challenge");
 }
 
-pub fn keygen() -> (Scalar, XOnly) {
+pub fn keygen() -> (Scalar, Point<EvenY>) {
     let mut x = Scalar::random(&mut thread_rng());
-    let X = XOnly::from_scalar_mul(G, &mut x);
+    let X = Point::even_y_from_scalar_mul(G, &mut x);
     (x, X)
 }
 
-pub fn sign(keypair: &(Scalar, XOnly), message: &[u8]) -> Signature {
+pub fn sign(keypair: &(Scalar, Point<EvenY>), message: &[u8]) -> Signature {
     let (x, X) = keypair;
     let mut r = Scalar::random(&mut thread_rng());
-    let R = XOnly::from_scalar_mul(G, &mut r);
+    let R = Point::even_y_from_scalar_mul(G, &mut r);
     let c = Scalar::from_hash(BIP340_CHALLENGE.clone().add(&R).add(X).add(message));
     let s = s!(r + c * x);
 
@@ -37,14 +38,14 @@ pub fn sign(keypair: &(Scalar, XOnly), message: &[u8]) -> Signature {
     }
 }
 
-pub fn verify(public_key: &XOnly, message: &[u8], Signature { R, s }: &Signature) -> bool {
+pub fn verify(public_key: Point<EvenY>, message: &[u8], Signature { R, s }: &Signature) -> bool {
     let X = public_key;
     let c = Scalar::from_hash(BIP340_CHALLENGE.clone().add(R).add(X).add(message)).mark::<Public>();
-    g!(s * G - c * { X.to_point() }) == *R
+    g!(s * G - c * X) == *R
 }
 
 fn main() {
     let keypair = keygen();
     let signature = sign(&keypair, b"attack at dawn");
-    assert!(verify(&keypair.1, b"attack at dawn", &signature));
+    assert!(verify(keypair.1, b"attack at dawn", &signature));
 }